+++ checkoutmanager/utils.py
@@ -1,5 +1,6 @@
 from __future__ import print_function
-from cStringIO import StringIO
+from __future__ import unicode_literals
+from six.moves import cStringIO
 from functools import wraps
 import os
 import subprocess
@@ -49,8 +50,11 @@ def system(command, input=None):
                          stdout=subprocess.PIPE,
                          stderr=subprocess.PIPE,
                          close_fds=MUST_CLOSE_FDS)
+    if input:
+        input = input.encode()
     stdoutdata, stderrdata = p.communicate(input=input)
     result = stdoutdata + stderrdata
+    result = result.decode()
     if p.returncode:
         raise CommandError(p.returncode, command, result)
 
@@ -66,7 +70,7 @@ def capture_stdout(func):
 
     @wraps(func)
     def newfunc(*args, **kwargs):
-        sys.stdout = StringIO()
+        sys.stdout = cStringIO()
         try:
             func(*args, **kwargs)
             return sys.stdout.getvalue()